requirements.txt
llama-cpp-python
TTS
optimum
openai
openai-whisper
SpeechRecognition
pyaudio
openwakeword
torch
scipy
numpy
transformers